Lambda

+++ m3-lambda/src/main/java/com/cydeo/Apple.java
@@ -14,7 +14,7 @@ public class Apple {
 
     private Color appleColor;
 
-    private Double appleWeight;
+    private int appleWeight;
 
 
 }
